Functor superclass; evalAttemptT; removed success.
Success should always be just a synonym for return, so why have it?

It was getting annoying not being able to call fmap where ever I wanted.
I'd like to make Applicative a superclass too, but we know some people
don't define their Monads properly.

evalAttemptT is something I've wanted, it just took a while to come up
with a name. Thought about drop, since it's really the opposite of lift.