Reworked the last changes for better extensible support.
failure calls SomeException, not toException. Otherwise, failure is
implicitly changing the exception that users throw.

Removed the extra cast operation in makeHandler. The test passes without
it, and I don't know why it's necesary. If I see an example that
requires it, I'll be happy to put it back in.